+++ R/s3.methods.R
@@ -355,15 +355,20 @@ hypothesis.brmsfit <- function(x, hypothesis, class = "b", alpha = 0.05, ...) {
     if (!all(parsH %in% pars)) 
       stop(paste("The following parameters cannot be found in the model:", 
                  paste0(gsub("__", ":", parsH[which(!parsH %in% pars)]), collapse = ", ")))
-    samples <- posterior.samples(x, parameters = rename(parsH, "__", ":"), fixed = TRUE)
-    names(samples) <- rename(varsH, c("[", "]"), c("OB", "CB"))
+    samples <- structure(posterior.samples(x, parameters = rename(parsH, "__", ":"), fixed = TRUE),
+                         names = rename(varsH, c("[", "]"), c("OB", "CB")))
+    samples <- matrix(with(samples, eval(parse(text = rename(h, c("[", "]"), c("OB", "CB"))))), ncol=1)
+    prior_samples <- get_prior_samples(x, pars = rename(parsH, "__", ":"))
+    if (!is.null(prior_samples)) {
+      names(prior_samples) <- rename(varsH, c("[", "]"), c("OB", "CB"))
+      prior_samples <- matrix(with(prior_samples, eval(parse(text = rename(h, c("[", "]"), c("OB", "CB"))))), ncol=1)
+    } 
     
     #evaluate hypothesis
-    out <- matrix(with(samples, eval(parse(text = rename(h, c("[", "]"), c("OB", "CB"))))), ncol=1)
     wsign <- ifelse(sign == "=", "equal", ifelse(sign == "<", "less", "greater"))
     probs <- switch(wsign, equal = c(alpha/2, 1-alpha/2), less = c(0, 1-alpha), greater = c(alpha, 1))
     out <- as.data.frame(matrix(unlist(lapply(c("mean","sd","quantile", "eratio"), get.estimate, 
-                         samples = out, probs = probs, hypothesis = wsign)), nrow = 1))
+             samples = samples, probs = probs, wsign = wsign, prior_samples = prior_samples)), nrow = 1))
     if (sign == "<") out[1,3] <- -Inf
     else if (sign == ">") out[1,4] <- Inf
     out <- cbind(out, ifelse(!(out[1,3] <= 0 && 0 <= out[1,4]), '*', ''))