+++ News.md
@@ -6,6 +6,8 @@ output: html_document
 # under development
 ----------------------------------------------------------------
 
+* allow for sampling from all specified proper priors in the model
+* calculate Bayes factors for point hypotheses in S3 method 'hypothesis'
 * fix a bug that could cause an error for weighted poisson and exponential models
 
 # brms 0.4.0